adds feature test for contacts
