+++ app/views/comments/_form.html.erb
@@ -6,6 +6,6 @@
     <%=f.text_field :content %>
   </p>
   <p>
-    <%= f.submit "Submit" %>
+    <%= f.submit %>
   </p>
 <% end %>
